refactor: move PaddleOCR component to Baidu category
Place the PaddleOCR component under the existing Baidu provider category
instead of Files & Knowledge, matching the repository's third-party provider
component organization.
