+++ src/lfx/src/lfx/components/files_and_knowledge/__init__.py
@@ -11,7 +11,6 @@ if TYPE_CHECKING:
     from lfx.components.files_and_knowledge.ingestion import KnowledgeIngestionComponent
     from lfx.components.files_and_knowledge.knowledge import KnowledgeComponent
     from lfx.components.files_and_knowledge.memory_retrieval import MemoryBaseComponent
-    from lfx.components.files_and_knowledge.paddleocr import PaddleOCRComponent
     from lfx.components.files_and_knowledge.retrieval import KnowledgeBaseComponent
     from lfx.components.files_and_knowledge.save_file import SaveToFileComponent
 
@@ -24,7 +23,6 @@ _dynamic_imports = {
     "KnowledgeIngestionComponent": "ingestion",
     "KnowledgeBaseComponent": "retrieval",
     "MemoryBaseComponent": "memory_retrieval",
-    "PaddleOCRComponent": "paddleocr",
     "SaveToFileComponent": "save_file",
 }
 
@@ -36,7 +34,6 @@ __all__ = [
     "KnowledgeComponent",
     "KnowledgeIngestionComponent",
     "MemoryBaseComponent",
-    "PaddleOCRComponent",
     "SaveToFileComponent",
 ]
